split us/hd on the first func
