+++ config/splat.hd.dra.yaml
@@ -201,6 +201,7 @@ segments:
       - [0x71574, c, 71830]
       - [0x7291C, c, 72BB0]
       - [0x75C44, c, 75F54]
+      - [0x7848C, c, 7879C]
       - [0x78A04, c, 78D0C]
       - [0x7A1C8, c, 7A4D0]
       - [0x7E168, c, 7E4BC]
