+++ config/splat.us.dra.yaml
@@ -266,6 +266,7 @@ segments:
       - [0x71830, c, 71830]
       - [0x72BB0, c, 72BB0]
       - [0x75F54, c, 75F54]
+      - [0x7879C, c, 7879C]
       - [0x78D0C, c, 78D0C]
       - [0x7A4D0, c, 7A4D0]
       - [0x7E4BC, c, 7E4BC]
